Added Motor Current Test Project
Created a new project to read the current and velocity of the motor in order to determine the change in current after a force exerted on the motor. The purpose of this project is to determine the correlation between the current and the force to determine how much force is exerted on a game piece using the current if we were to use some sort of gripper mechanism that applies pressure to the game pieces. This would allow us to determine when to stop the mechanism, not harming the game piece while still holding it tight.